Enforce cargo clippy check in CI

+++ crates/gpu-test-arith-basic/src/lib.rs
@@ -6,6 +6,6 @@
 #[gpu_codegen::kernel]
 pub fn kernel_arith(a: &[u8], b: &mut [u8]) {
     gpu::add_mlir_string_attr("#gpu<dim x>");
-    let c = gpu::thread_id() as usize;
+    let c = gpu::thread_id();
     b[c] = a[c];
 }
